Large DDGI progress update
Successfully implemented an improved raymarching method for DDGI. Previously was raymarching into the gbuffer which was simply an inferior technique with poor approximations due to screen space limitations. The new method voxelizes the whole world in a 3d texture, then ray marches into that 3d texture to determine ray hits. To generate the 3d texture a voxelize.compute shader was devised to determine voxel occupancy by testing scene triangle collision with the voxels. This required building a SSBO of scene triangles to send to the shader.